Mac: UI tweaks for task manager.
Nib change: Reduce row height, make scrollbars smaller.

Make text in task manager table slightly smaller.

Show in decimal digit for %cpu.

Show memory in KB/MB, not always in K.

Change update frequency from 1s to 2s to match Activity Monitor's default.

(all mac-only. ui team is fine with this.)

Finally, turn taskman on.

BUG=13156
TEST=Open task manager, look at it. Should look & feel similar to Activity Monitor.

Review URL: http://codereview.chromium.org/536038

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@36096 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/process_util.h
@@ -429,10 +429,10 @@ class ProcessMetrics {
   // Returns the CPU usage in percent since the last time this method was
   // called. The first time this method is called it returns 0 and will return
   // the actual CPU info on subsequent calls.
-  // Note that on multi-processor machines, the CPU usage value is for all
-  // CPUs. So if you have 2 CPUs and your process is using all the cycles
-  // of 1 CPU and not the other CPU, this method returns 50.
-  int GetCPUUsage();
+  // On Windows, the CPU usage value is for all CPUs. So if you have 2 CPUs and
+  // your process is using all the cycles of 1 CPU and not the other CPU, this
+  // method returns 50.
+  double GetCPUUsage();
 
   // Retrieves accounting information for all I/O operations performed by the
   // process.
